Fix typing error in secrets package and organize its adapters into a separate subdirectory.
